test: adding matrix 4 tests and looking for a bug in submatrix

+++ src/libla/libla.h
@@ -52,35 +52,35 @@ enum	e_vec4_index
 // --- Structure : Vector ---
 typedef struct s_vec4
 {
-	float	data[4];
+	double	data[4];
 }	t_vec4;
 
 typedef struct s_vec3
 {
-	float	data[3];
+	double	data[3];
 }	t_vec3;
 
 // --- Structure : Matrix ---
 typedef struct s_mat4
 {
-	float	data[4][4];
+	double	data[4][4];
 }	t_mat4;
 
 typedef struct s_mat3
 {
-	float	data[3][3];
+	double	data[3][3];
 }	t_mat3;
 
 typedef struct s_mat2
 {
-	float	data[2][2];
+	double	data[2][2];
 }	t_mat2;
 
 // =============================================================================
 // Section : General
 // =============================================================================
 
-bool	equalf(float a, float b);
+bool	equalf(double a, double b);
 
 // =============================================================================
 // Section : Vector
@@ -102,23 +102,23 @@ t_vec4	vec4_sub(const t_vec4 *a, const t_vec4 *b);
 
 // --- Multiplication ---
 t_vec3	vec3_hadamard_product(const t_vec3 *v1, const t_vec3 *v2);
-t_vec3	vec3_mul(const t_vec3 *v, float scalar);
-t_vec4	vec4_mul(const t_vec4 *v, float scalar);
+t_vec3	vec3_mul(const t_vec3 *v, double scalar);
+t_vec4	vec4_mul(const t_vec4 *v, double scalar);
 
 // --- Division ---
-t_vec4	vec4_div(const t_vec4 *v, float scalar);
+t_vec4	vec4_div(const t_vec4 *v, double scalar);
 
 // --- Inversion ---
 t_vec4	vec4_inv(const t_vec4 *v);
 
 // --- Magnitude ---
-float	vec4_magnitude(const t_vec4 *v);
+double	vec4_magnitude(const t_vec4 *v);
 
 // --- Normalization ---
 t_vec4	vec4_normalize(const t_vec4 *v);
 
 // --- Dot Product ---
-float	vec4_dot_product(const t_vec4 *v1, const t_vec4 *v2);
+double	vec4_dot_product(const t_vec4 *v1, const t_vec4 *v2);
 
 // --- Cross Product ---
 t_vec4	vec4h_cross_product(const t_vec4 *v1, const t_vec4 *v2);
@@ -139,28 +139,28 @@ t_mat3	mat4_submatrix(const t_mat4 *m, short row, short col);
 t_mat2	mat3_submatrix(const t_mat3 *m, short row, short col);
 
 // --- Determinant ---
-float	mat2_det(const t_mat2 *a);
-float	mat3_det(const t_mat3 *a);
-float	mat4_det(const t_mat4 *a);
+double	mat2_det(const t_mat2 *a);
+double	mat3_det(const t_mat3 *a);
+double	mat4_det(const t_mat4 *a);
 
 // --- Inversion ---
 t_mat4	mat4_inv(const t_mat4 *a);
 
 // --- Minor ---
-float	mat3_minor(const t_mat3 *m, short row, short col);
-float	mat4_minor(const t_mat4 *m, short row, short col);
+double	mat3_minor(const t_mat3 *m, short row, short col);
+double	mat4_minor(const t_mat4 *m, short row, short col);
 
 // --- Cofactor ---
-float	mat4_cofactor(const t_mat4 *a, short row, short col);
-float	mat3_cofactor(const t_mat3 *a, short row, short col);
+double	mat4_cofactor(const t_mat4 *a, short row, short col);
+double	mat3_cofactor(const t_mat3 *a, short row, short col);
 
 // --- Multiplication ---
-t_mat4	mat4_mul(const t_mat4 *a, float scalar);
+t_mat4	mat4_mul(const t_mat4 *a, double scalar);
 t_mat4	mat4_mat4_mul(const t_mat4 *a, const t_mat4 *b);
 t_vec4	mat4_vec4_mul(const t_mat4 *a, const t_vec4 *b);
 
 // --- Division ---
-t_mat4	mat4_div(const t_mat4 *a, float scalar);
+t_mat4	mat4_div(const t_mat4 *a, double scalar);
 
 // --- Transpose ---
 t_mat4	mat4_transpose(const t_mat4 *a);
