+++ src/libla/matrix/determinant/mat2_det.c
@@ -12,7 +12,7 @@
 
 #include "libla.h"
 
-float	mat2_det(const t_mat2 *a)
+double	mat2_det(const t_mat2 *a)
 {
 	return (a->data[0][0] * a->data[1][1] - \
 			a->data[0][1] * a->data[1][0]);
